Add Jest tests for backend
