+++ Talentify-backend/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "CI=true jest",
     "start": "node server.js"
   },
   "keywords": [],
@@ -20,5 +20,13 @@
     "express-rate-limit": "^7.5.1",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.16.0"
+  },
+  "devDependencies": {
+    "jest": "^30.0.3",
+    "mongodb-memory-server": "^10.1.4",
+    "supertest": "^7.1.1"
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
